new zarr loader is ready! sub it in
